// src/app/page/master/map-event-to-produk/map-event-to-produk.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapEventToProduk {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}